Feat: keyword 상세 조회 API 구현
