+++ routes/keywordRoute.js
@@ -3,6 +3,5 @@ const router = express.Router();
 const keywordController = require("../controllers/keywordController");
 
 router.post("/", keywordController.create);
-router.get("/:keywordId", keywordController.list);
 
 module.exports = router;
